+++ changelog.md
@@ -2,9 +2,15 @@
 
 ## 1.0.0-alpha.1
 
+- **New:**
+- Preferences:
+  - `Enable tab completion` Enable tab complete, when enabled pressing tab will complete the current word form the last lines of buffered text
+  - `Tab completion buffer limit` The number of buffered lines to use for tab completion
+  - `Ignore case for tab completion` Ignore letter cashing when searching for matching words
 - **Changed:**
   - Display: Adjust word wrap break points to have better flow
   - Update electron 19.1.5 to 19.1.6
+  - Update mathjs 11.3.3 to 11.4.0
 
 ## 1.0.0-alpha